[data-spike-sitrep] - Refactored to use ArgumentParser and stdout - TT
